+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 50
 * Use field in basic_fields and call sites
 * Use field in basic_parser
 * Tidy up basic_fields, header, and field concepts
+* Fields concept work
 
 API Changes: